掉落物移到EntityZombie
