+++ src/main/java/cn/nukkit/entity/mob/EntityHusk.java
@@ -17,15 +17,12 @@ import cn.nukkit.entity.ai.route.finder.impl.SimpleFlatAStarRouteFinder;
 import cn.nukkit.entity.ai.route.posevaluator.WalkingPosEvaluator;
 import cn.nukkit.entity.ai.sensor.NearestPlayerSensor;
 import cn.nukkit.entity.data.IntEntityData;
-import cn.nukkit.item.Item;
 import cn.nukkit.level.format.FullChunk;
 import cn.nukkit.nbt.tag.CompoundTag;
 import cn.nukkit.network.protocol.LevelSoundEventPacket;
 import cn.nukkit.potion.Effect;
-import cn.nukkit.utils.Utils;
 
 import java.util.Set;
-import java.util.concurrent.ThreadLocalRandom;
 
 /**
  * @author PikyCZ
@@ -110,17 +107,4 @@ public class EntityHusk extends EntityZombie {
         return true;
     }
 
-    //使用dorps判断概率,在0.83%概率下会给土豆或者马铃薯以及铁锭任意一个物品
-    @Override
-    public Item[] getDrops() {
-        float drops = ThreadLocalRandom.current().nextFloat(100);
-        if (drops < 0.83) {
-            return switch (Utils.rand(0, 2)) {
-                case 0 -> new Item[]{Item.get(Item.IRON_INGOT, 0, 1), Item.get(Item.ROTTEN_FLESH, 0, 1)};
-                case 1 -> new Item[]{Item.get(Item.CARROT, 0, 1), Item.get(Item.ROTTEN_FLESH, 0, 1)};
-                default -> new Item[]{Item.get(Item.POTATO, 0, 1), Item.get(Item.ROTTEN_FLESH, 0, 1)};
-            };
-        }
-        return new Item[]{Item.get(Item.ROTTEN_FLESH, 0, Utils.rand(0, 2))};
-    }
 }
